Move BasePoolMethodsV3 to poolsV3 folder
