+++ modules/sor/sorV2/lib/poolsV2/basePool.ts
@@ -20,8 +20,3 @@ export interface BasePool {
      */
     getPoolTokens(tokenIn: Token, tokenOut: Token): { tIn: BasePoolToken; tOut: BasePoolToken };
 }
-
-export interface BasePoolMethodsV3 extends BasePool {
-    tokens: (BasePoolToken | Erc4626PoolToken)[];
-    getPoolState(): PoolState | BufferState;
-}
