+++ modules/sor/sorV2/lib/poolsV3/basePoolMethodsV3.ts
@@ -1,0 +1,9 @@
+import { BufferState } from '@balancer-labs/balancer-maths';
+import { PoolState } from '@balancer-labs/balancer-maths';
+import { BasePool } from '../poolsV2/basePool';
+import { BasePoolToken, Erc4626PoolToken } from '../utils';
+
+export interface BasePoolMethodsV3 extends BasePool {
+    tokens: (BasePoolToken | Erc4626PoolToken)[];
+    getPoolState(): PoolState | BufferState;
+}
